// apps/web/src/components/v2/utils/tabUtils/link.type.ts
/*
 * Copyright (c) 2024. Devtron Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export interface iLink {
    id?: number
    name?: string
    icon?: string
    className?: string
    isSelected: boolean
    isDisabled?: boolean
    isHidden?: boolean
    status?: string
    count?: number
    url?: string
    isDeleted?: boolean
    uid?: string
}

export interface iLinks extends Readonly<Array<iLink>> {}